+++ docs/CHANGELOG.md
@@ -34,7 +34,13 @@ const result = await management.bitcoin_get_balance_query({
 
 ### Changed
 
+* feat!: support for restricting II auth methods
+    * New login option: `allowPinAuthentication?: boolean;`
+    * Response from II includes `authnMethod: 'passkey' | 'pin' | 'recovery';`
+    * OnSuccess now optionally passes the message directly from the IDP provider
+    * Support for arbitrary login values passed to IDP through `customValues` option
 * fix: pads date numbers in changelog automation. E.G. 2024-3-1 -> 2024-03-01
+
 * feat: allow passing `DBCreateOptions` to `IdbStorage` constructor
 * updated management canister interface